renaming common package
