+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/common/MonthSelectionTopBar.kt
@@ -18,7 +18,7 @@
  *
  */
 
-package br.com.rbrthmn.ui.financialcompanion.commoncomponents
+package br.com.rbrthmn.ui.financialcompanion.common
 
 import InfiniteHorizontalCircularList
 import InfiniteVerticalCircularList
