+++ app/src/main/java/br/com/rbrthmn/ui/financialcompanion/screens/home/components/CreditCardBillsCard.kt
@@ -57,7 +57,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import br.com.rbrthmn.R
-import br.com.rbrthmn.ui.financialcompanion.commoncomponents.DecimalInputField
+import br.com.rbrthmn.ui.financialcompanion.common.DecimalInputField
 import org.koin.androidx.compose.koinViewModel
 
 @Composable
